feat(workflow): Lazily load user code into Workflow isolate
